requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0